seeded the database with geocoder and change schema to work with geocode

+++ db/migrate/20151006174356_create_greenthumbs.rb
@@ -5,8 +5,8 @@ class CreateGreenthumbs < ActiveRecord::Migration
       t.text :address
       t.text :boro
       t.decimal :size, :precision => 3, :scale => 3
-      t.decimal :longitude
-      t.decimal :latitude
+      t.float :longitude
+      t.float :latitude
     end
   end
 end
